refactor(diagnostic): extract shared Severity/Finding to hecks_life/src/diagnostic.rs (i59)
Three validators (duplicate_policy, lifecycle, io) independently
defined the same Severity enum + Finding struct + impls. Extracting
to a shared module removes the duplication and unblocks Phase B
specialization — once the boilerplate is shared, each validator's
shape only needs to model the rule, not the ceremony.

New module hecks_life/src/diagnostic.rs (~63 LoC):
  - Severity { Error, Warning } — union of all three variants
    (duplicate_policy is Error-only today but Severity now has both;
    it just doesn't emit Warning findings)
  - Finding struct (severity, location, message)
  - Finding::err / Finding::warn constructors (public)
  - Finding::icon() — dispatches on severity

Each validator:
  - pub use crate::diagnostic::{Finding, Severity} — keeps integration
    tests' existing imports working; no test changes needed
  - Local Severity + Finding + Finding impl blocks deleted
  - Report + Report impls stay per-validator (shapes differ: flat
    Vec<Finding> vs static/runtime partition; future refactor)

Size: +63 (diagnostic.rs) / -56 (duplicated boilerplate removed).
Net +7 LoC but three sources of truth collapsed to one.

What this unblocks:
  - duplicate_policy_validator retirement — ~25 LoC of actual rule
    logic worth specializing, instead of 95 LoC of mostly ceremony
  - lifecycle_validator retirement
  - io_validator retirement

Tests:
  - cargo test --release: all 26 suites green
  - Parity: 499/506 (unchanged)

[antibody-exempt: hecks_life/src/diagnostic.rs is a new Rust module
 that collapses duplicated validator infrastructure (Severity enum
 + Finding struct) into one shared definition. Non-DSL; pure Rust
 refactor required as prerequisite per i59 for the next wave of
 i51 Phase B retirements.]

[antibody-exempt: hecks_life/src/lib.rs adds pub mod diagnostic
 registration for the new shared module.]

[antibody-exempt: hecks_life/src/duplicate_policy_validator.rs,
 io_validator.rs, lifecycle_validator.rs — replace locally-defined
 Severity + Finding + impls with pub use of crate::diagnostic.
 Same three files the i59 inbox item explicitly names as blocked
 on this extraction.]
